Last modif before presentation

+++ src/application/essivi/models/type_Vente.py
@@ -10,7 +10,7 @@ class Type_vente(db.Model):
     libelle_type_vente = db.Column(db.String(20), nullable=False)
     prix_unit = db.Column(db.Integer, nullable=False)
     image = db.Column(db.String(150), nullable=True)
-    qte_contenu_unitaire = db.Column(db.Integer, nullable=False)
+    qte_contenu_unitaire = db.Column(db.Float, nullable=False)
     qte_composition = db.Column(db.Integer, default=1)
     marque_id = db.Column(db.Integer, db.ForeignKey('marques.id'), nullable=False)
